phoenix_rag_demo/requirements.txt
arize-phoenix
arize-phoenix-client
arize-phoenix-evals
arize-phoenix-otel

langchain
langchain-community
langchain-chroma
langchain-huggingface

sentence-transformers
chromadb

google-generativeai
python-dotenv
pandas